pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.67.0",
    "pillow>=11.1.0",
    "pypdf2>=3.0.1",
    "reportlab>=4.3.1",
    "rich>=13.9.4",
]
